+++ src/main/java/racingcar/dto/ResultDto.java
@@ -3,9 +3,9 @@ package racingcar.dto;
 // TODO: record로 변경 가능한 듯 함. 뭔지 자료조사하고, 적용가능하면 적용해볼 것
 public class ResultDto {
     private final String name;
-    private final Long totalMoveCount;
+    private final int totalMoveCount;
 
-    public ResultDto(String name, Long totalMoveCount) {
+    public ResultDto(String name, int totalMoveCount) {
         this.name = name;
         this.totalMoveCount = totalMoveCount;
     }
@@ -14,7 +14,7 @@ public class ResultDto {
         return name;
     }
 
-    public Long getTotalMoveCount() {
+    public int getTotalMoveCount() {
         return totalMoveCount;
     }
